Move runpaths out of shared
