+++ tests/unit_tests/c_wrappers/res/enkf/test_ert_run_context.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 
 from ert._c_wrappers.enkf import RunContext
-from ert.shared.runpaths import Runpaths
+from ert.runpaths import Runpaths
 
 
 @pytest.mark.usefixtures("use_tmpdir")
